feat(pie-monorepo): separated code into files
