+++ apps/examples/wc-next10/Counter.js
@@ -1,6 +1,6 @@
 import React, { useState, useRef } from 'react';
 import { createComponent } from '@lit-labs/react';
-import { PieButton, BUTTON_SIZE, BUTTON_VARIANT } from '@justeattakeaway/pie-button';
+import { PieButton, BUTTON_SIZE } from '@justeattakeaway/pie-button';
 
 const PieBtn = createComponent({
     tagName: 'pie-button',
@@ -10,24 +10,13 @@ const PieBtn = createComponent({
 });
 
 export default function Counter () {
-    const variantIndex = useRef(0)
     const [count, setCount] = useState(0);
-    const [variantName, setVariantName] = useState(BUTTON_VARIANT.PRIMARY);
 
     const onIncrement = () => setCount(count + 1);
     const onDecrement = () => setCount(count - 1);
-    const switchVariant = () => {
-        variantIndex.current += 1;
-        const variantsKeys = Object.keys(BUTTON_VARIANT);
-        const variantKey = variantsKeys[variantIndex.current % variantsKeys.length];
-        const variant = BUTTON_VARIANT[variantKey];
-
-        setVariantName(variant);
-    };
 
     return (
         <>
-            <h2>Test click event</h2>
             <div>
                 <h3>Counter</h3>
                 <div className="flex-wrapper">
@@ -36,31 +25,6 @@ export default function Counter () {
                         Counter: { count }
                     </div>
                     <PieBtn onClick={onIncrement}>increment</PieBtn>
-                </div>
-            </div>
-            <h2>Test props</h2>
-            <div>
-                <h3>Click the button to switch the variant</h3>
-                <div className="flex-wrapper">
-                    <PieBtn variant={variantName} onClick={switchVariant}>
-                        Switch variant
-                    </PieBtn>
-                    <div className='padding'>Variant: <b>{variantName}</b></div>
-                </div>
-            </div>
-            <div>
-                <h3>Button sizes</h3>
-                <div>
-                    {
-                        Object.keys(BUTTON_SIZE)
-                            .map((key) => {
-                                const size= BUTTON_SIZE[key];
-
-                                return (
-                                    <PieBtn size={size} key={key}>{key.toLowerCase()}</PieBtn>
-                                );
-                            })
-                    }
                 </div>
             </div>
         </>
