+++ apps/examples/wc-next10/pages/index.jsx
@@ -2,12 +2,16 @@ import React from 'react';
 import dynamic from 'next/dynamic'
 
 const Counter = dynamic(() => import('../Counter'), { ssr: false });
+const VariantSwitch = dynamic(() => import('../VariantSwitch'), { ssr: false });
+const ButtonSizes = dynamic(() => import('../ButtonSizes'), { ssr: false });
 
 export default function Home() {
     return (
         <>
             <main>
                 <Counter />
+                <VariantSwitch />
+                <ButtonSizes />
             </main>
         </>
     )
